Initial End to End Test (#2)
This adds an initial end to end test that will use an "outside" app to send
a profile to pprof me, and then request a web view of that profile and
ensure that we get back some for of expected html.

An initial test in the app directory for the REST endpoints was also added
to test the added GET /profiles endpoint. One change was introduced
in order to make those tests work better:
> One change here is to allow the App to listen on a server
assigned port (i.e. pass in port 0 in the LISTEN_ADDR). This resulted
in a change to app.New() to additionally return a error interface.